chatbot_project/requirements.txt
aiofiles==22.1.0
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.18.0
annotated-types==0.7.0
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.3
attrs==22.2.0
Automat==22.10.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.2
biclustlib==0.0.13
bicon==1.3.4
biopython==1.81
bioservices==1.11.2
biothings-client==0.3.0
bleach==6.0.0
Bottleneck==1.3.7
cattrs==23.1.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
colorlog==6.7.0
comm==0.1.2
constantly==15.1.0
contourpy==1.1.0
cryptography==40.0.2
cssselect==1.2.0
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==4.1.7
django-cors-headers==4.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.3.0
docopt==0.6.2
easydev==0.12.1
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.12.0
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2024.6.1
gevent==23.9.1
goatools==1.3.1
greenlet==3.0.0rc3
grequests==0.7.0
gseapy==0.9.15
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.2
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.19.0
kiwisolver==1.4.5
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==2.0.5
mpmath==1.3.0
multidict==6.0.4
munkres==1.1.4
mygene==3.2.2
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
openai==1.37.0
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parsel==1.8.1
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.37
Protego==0.2.1
psutil==5.9.4
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
PyDispatcher==2.0.7
pydot==1.4.2
Pygments==2.14.0
PyJWT==2.8.0
pymongo==4.3.3
pyOpenSSL==23.1.1
pyparsing==3.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
queuelib==1.6.2
regex==2024.7.24
requests==2.28.2
requests-cache==1.1.0
requests-file==1.5.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpy2==3.5.14
safetensors==0.4.4
scikit-learn==1.4.2
scipy==1.11.2
Scrapy==2.8.0
seaborn==0.12.2
Send2Trash==1.8.0
sentencepiece==0.2.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.2.4
stack-data==0.6.2
statsmodels==0.14.0
suds-community==1.1.2
sympy==1.13.2
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tldextract==3.4.0
tokenizers==0.19.1
torch==2.2.2
torch-utils==0.1.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.44.0
Twisted==22.10.0
typing_extensions==4.12.2
tzlocal==5.0.1
uri-template==1.2.0
url-normalize==1.4.3
urllib3==1.26.14
w3lib==2.1.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
wrapt==1.15.0
XlsxWriter==3.1.3
xmltodict==0.13.0
y-py==0.5.9
yarl==1.9.2
ypy-websocket==0.8.2
zope.event==5.0
zope.interface==6.0
